#244512 purchase: fix ratio, type item count, fix related prop rest, fix refresh basket

+++ lib/trading/action/incoming/item.php
@@ -23,7 +23,7 @@ class Item extends Skeleton
 		return (int)$this->requireField('id');
 	}
 
-	public function getQuantity() : int
+	public function getQuantity() : float
 	{
 		return (float)$this->requireField('count');
 	}
